t/Test/Mini/Unit/Logger/Lapidary.t passes
Well, kinda.  There's a small bug in the tests
that causes them to fail if invoked from anywhere
but the root of the checkout, but they do issue
a proper exit code if run from there.

At this point, all current test suites pass.
Anything, therefore, that doesn't presently work
is due to an oversight in the universe at large.
Normality will be returned to just as soon as we
can once again agree upon just what normality is.